backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.11.1; python_version >= '3.9'
boto3==1.42.95; python_version >= '3.9'
botocore==1.42.95; python_version >= '3.9'
django==5.2.13; python_version >= '3.10'
django-cors-headers==4.9.0; python_version >= '3.9'
django-storages[s3]==1.14.6; python_version >= '3.7'
djangorestframework==3.17.1; python_version >= '3.10'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
gunicorn==25.3.0; python_version >= '3.10'
jmespath==1.1.0; python_version >= '3.9'
packaging==26.1; python_version >= '3.8'
psycopg2-binary==2.9.12; python_version >= '3.9'
pyjwt==2.12.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-decouple==3.8
s3transfer==0.16.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.5; python_version >= '3.8'
urllib3==2.6.3; python_version >= '3.9'
